requirements.txt
langchain
langchain-community
openai
python-dotenv
tiktoken
langchain-ollama